refactor: 점수 관련 로직 ChessGame으로 이동
