+++ src/main/java/chess/game/ChessGame.java
@@ -5,20 +5,24 @@ import static java.util.stream.Collectors.toList;
 import chess.domain.Board;
 import chess.domain.BoardFactory;
 import chess.domain.Position;
+import chess.domain.Role;
 import chess.domain.Team;
+import chess.domain.piece.Piece;
 import chess.dto.SquareResponse;
 import chess.game.state.GameState;
-import chess.game.state.running.CheckedState;
 import chess.game.state.end.EndState;
 import chess.game.state.end.NoneWinState;
+import chess.game.state.running.CheckedState;
 import chess.game.state.waiting.WaitingState;
 import java.util.List;
+import java.util.Map;
 import java.util.function.BiConsumer;
 import java.util.function.Supplier;
 
 public class ChessGame {
     private static final String INVALID_TURN_EXCEPTION_MESSAGE = "[ERROR] 해당 팀의 턴이 아닙니다.";
     private static final String EMPTY_PIECE_EXCEPTION_MESSAGE = "[ERROR] 빈 칸은 움직일 수 없습니다.";
+    private static final double PAWN_COUNT_MULTIPLIER = 0.5;
 
     private Board board;
     private GameState gameState;
@@ -74,7 +78,34 @@ public class ChessGame {
     }
 
     public double getTeamScore(Team team) {
-        return gameState.getTeamScore(() -> board.getTeamScore(team).getScore());
+        return gameState.getTeamScore(() -> calculateScore(team));
+    }
+
+    private double calculateScore(Team team) {
+        double pawnScore = getPawnScore(team);
+        double withoutPawnScore = getWithoutPawnScore(team);
+        return pawnScore + withoutPawnScore;
+    }
+
+    private double getPawnScore(Team team) {
+        Map<Integer, Long> pawnCountByX = board.getPawnCountByX(team);
+        return pawnCountByX.values().stream()
+                .mapToDouble(this::calculatePawnScore)
+                .sum();
+    }
+
+    private double calculatePawnScore(long pawnCount) {
+        if (pawnCount == 1) {
+            return Role.PAWN.getScore();
+        }
+        return pawnCount * Role.PAWN.getScore() * PAWN_COUNT_MULTIPLIER;
+    }
+
+    private double getWithoutPawnScore(Team team) {
+        return board.getTeamSquares(team).values().stream()
+                .filter(piece -> !piece.isRoleOf(Role.PAWN))
+                .mapToDouble(Piece::getScore)
+                .sum();
     }
 
     public Team getTurn() {
